For #647: move integrate yaml test case's schema to api package
