+++ sharding-jdbc-core/src/test/java/io/shardingjdbc/core/integrate/api/yaml/YamlMasterSlaveIntegrateTest.java
@@ -15,14 +15,13 @@
  * </p>
  */
 
-package io.shardingjdbc.core.integrate.api.yaml.masterslave;
+package io.shardingjdbc.core.integrate.api.yaml;
 
 import com.google.common.base.Function;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
 import io.shardingjdbc.core.api.ConfigMapContext;
 import io.shardingjdbc.core.api.MasterSlaveDataSourceFactory;
-import io.shardingjdbc.core.integrate.api.yaml.AbstractYamlDataSourceTest;
 import lombok.RequiredArgsConstructor;
 import org.junit.Test;
 import org.junit.runner.RunWith;
