+++ sharding-jdbc-core/src/test/java/io/shardingjdbc/core/integrate/api/yaml/YamlShardingWithMasterSlaveIntegrateTest.java
@@ -15,13 +15,12 @@
  * </p>
  */
 
-package io.shardingjdbc.core.integrate.api.yaml.sharding;
+package io.shardingjdbc.core.integrate.api.yaml;
 
 import com.google.common.base.Function;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
 import io.shardingjdbc.core.api.ShardingDataSourceFactory;
-import io.shardingjdbc.core.integrate.api.yaml.AbstractYamlDataSourceTest;
 import lombok.RequiredArgsConstructor;
 import org.junit.Test;
 import org.junit.runner.RunWith;
